Refactor emojis into its own module
